+++ src/main/java/user/database/UserResetPasswordDatabaseGateway.java
@@ -1,0 +1,7 @@
+package user.database;
+
+import user.entities.User;
+
+public interface UserResetPasswordDatabaseGateway {
+    User ResetPassword(String userName, String passWord);
+}
